move the learning tracks processing into a separate module and return featured link in addition to learning tracks
