+++ lib/page.js
@@ -11,6 +11,7 @@ const pathUtils = require('./path-utils')
 const Permalink = require('./permalink')
 const languages = require('./languages')
 const renderContent = require('./render-content')
+const processLearningTracks = require('./process-learning-tracks')
 const { renderReact } = require('./react/engine')
 const { productMap } = require('./all-products')
 const slash = require('slash')
@@ -207,23 +208,10 @@ class Page {
       this.permissions = await renderContent(this.rawPermissions, context)
     }
 
+    // Learning tracks may contain Liquid and need to have versioning processed.
     if (this.learningTracks) {
-      const learningTracks = []
-      for await (const rawTrackName of this.rawLearningTracks) {
-        // Track names in frontmatter may include Liquid conditionals
-        const renderedTrackName = await renderContent(rawTrackName, context, { textOnly: true, encodeEntities: true })
-        if (!renderedTrackName) continue
-
-        const track = context.site.data['learning-tracks'][context.currentProduct][renderedTrackName]
-        if (!track) continue
-        learningTracks.push({
-          trackName: renderedTrackName,
-          title: await renderContent(track.title, context, { textOnly: true, encodeEntities: true }),
-          description: await renderContent(track.description, context, { textOnly: true, encodeEntities: true }),
-          guides: await getLinkData(track.guides, context)
-        })
-      }
-
+      const { featuredTrack, learningTracks } = await processLearningTracks(this.rawLearningTracks, context)
+      this.featuredTrack = featuredTrack
       this.learningTracks = learningTracks
     }
